+++ core/pgspec/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "plasm-pgspec"
+name = "plasm-parent-pgspec"
 version = "0.1.0"
 authors = ["Takumi Yamashita <takumi@staked.co.jp>"]
 edition = "2018"
@@ -7,61 +7,61 @@ edition = "2018"
 [dependencies.parity-codec]
 default-features = false
 features = ['derive']
-version = '3.2'
+version = "4.1.1"
 
 [dependencies.primitives]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
 package = 'substrate-primitives'
-tag = 'v1.0.0rc2'
+version = '2.0.0'
 
 [dependencies.sr_primitives]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
 package = 'sr-primitives'
-tag = 'v1.0.0rc2'
+version = '2.0.0'
 
 [dependencies.sr-io]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
 package = 'sr-io'
-tag = 'v1.0.0rc2'
+version = '2.0.0'
 
 [dependencies.rstd]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
 package = 'sr-std'
-tag = 'v1.0.0rc2'
+version = '2.0.0'
 
 [dependencies.balances]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
 package = 'srml-balances'
-tag = 'v1.0.0rc2'
+version = '2.0.0'
 
 [dependencies.support]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
 package = 'srml-support'
-tag = 'v1.0.0rc2'
+version = '2.0.0'
 
 [dependencies.system]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
 package = 'srml-system'
-tag = 'v1.0.0rc2'
+version = '2.0.0'
 
 [dependencies.timestamp]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
 package = 'srml-timestamp'
-tag = 'v1.0.0rc2'
+version = '2.0.0'
 
 [dependencies.contract]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
-package = 'srml-contract'
-tag = 'v1.0.0rc2'
+package = 'srml-contracts'
+version = '2.0.0'
 
 [dependencies.serde]
 optional = true
